requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
httpx==0.25.2
requests==2.31.0
python-dotenv==1.0.0
motor==3.6.0
pymongo==4.9.2
pydantic==2.5.0